requirements.txt
behave-restful
bolt-ta
conttest
flask
nose

#added
Flask-SQLAlchemy
black
pylint
ipython
httpie